+++ CHANGELOG.md
@@ -128,6 +128,7 @@
 - [9460](https://github.com/vegaprotocol/vega/issues/9460) - Add APIs for volume discount program.
 - [9628](https://github.com/vegaprotocol/vega/issues/9628) - Upgrade `CometBFT`.
 - [9577](https://github.com/vegaprotocol/vega/issues/9577) - Feature test coverage for distribution strategy
+- [9542](https://github.com/vegaprotocol/vega/issues/9542) - Ensure all per asset accounts exist after migration
 
 ### 🐛 Fixes